Move portal notice url parser to ThePostDetail.vue
